Factored out heuristic computions out of Planner file
