+++ .gitignore
@@ -9,3 +9,4 @@ Interpreter.js
 Astar.js
 
 Experiment.js
+Heuristics.js
