+++ Makefile
@@ -27,6 +27,6 @@ grammar.js: grammar.ne
 run: offline
 	# node shrdlite-offline.js small "grasp the yellow box"
 	# node shrdlite-offline.js small "put the blue table on a box"
-	# node shrdlite-offline.js small "put the blue table on the red box"
+	node shrdlite-offline.js small "put the blue table on the red box"
 	# node shrdlite-offline.js small "put the white ball on the white ball"
-	node shrdlite-offline.js small "put the blue box in the red box"
+	# node shrdlite-offline.js small "put the blue box in the red box"
