+++ Planner.ts
@@ -1,6 +1,7 @@
 ///<reference path="World.ts"/>
 ///<reference path="Interpreter.ts"/>
 ///<reference path="Astar.ts"/>
+///<reference path="Heuristics.ts"/>
 
 module Planner {
 
@@ -47,7 +48,9 @@ module Planner {
         worldDictionary = state.objects;
 
         var goal = computeGoalFunction(intprt);
-        var heur = computeHeuristicFunction(intprt);
+        var heur = Heuristics.computeHeuristicFunction(intprt);
+
+        console.log("StartH: " + Heuristics.heuristicDifference(state, "g", "l", true) );
 
         var plan : string[] = Astar.astar(neighbours, cost, heur, state, goal, false, 10000);
         plan.shift();
@@ -57,207 +60,17 @@ module Planner {
         return plan;
     }
 
-    // Ducktyping subtype of WorldState :)
-    // should be sufficient.
-    interface State{
-        arm : number;
-        holding: string;
-        stacks : string[][];
-    }
-
-    function cost(a : State, b : State) : number{
+    function cost(a : Heuristics.State, b : Heuristics.State) : number{
         return 1;
     }
 
-    /**
-    * @return heuristic function for Astar.
-    */
-    function computeHeuristicFunction(intprt : Interpreter.Literal[][]) : Astar.Heuristic<State>{
-        return (s : State) => {
-            var hValue = Infinity;
-            for(var ix in intprt){
-                var hc = heuristicConjunctiveClause(s, intprt[ix]);
-                if(hValue > hc){
-                    hValue = hc;
-                }
-            }
-            // Return minimum heuristic value of the disjunction.
-            return hValue;
-        };
-    }
 
-    function heuristicConjunctiveClause(s : State, c : Interpreter.Literal[]) : number{
-        var hValue = 0;
-        for(var ix in c){
-            var hc = heuristicAtom(s, c[ix]);
-            if(hValue < hc){
-                hValue = hc;
-            }
-        }
-        // Return maximum heuristic value of the conjuction.
-        return hValue;
-    }
-
-    function heuristicAtom(s : State, atom : Interpreter.Literal) : number {
-
-        switch(atom.rel){
-            case "holding":
-                var target = atom.args[0];
-                var holds = s.holding === target;
-
-                if(atom.pol){
-                    return heuristicDistance(s, target);
-                }
-                if(holds){
-                    return 1;
-                    // Just drop it. Optionally: use canSupport function.
-                }
-                // Already done
-                return 0;
-
-            case "inside": // Same as ontop.
-            case "ontop":
-                var target = atom.args[0];
-                var below = atom.args[1];
-
-                if(atom.pol){
-                    // return 0;
-                    // return heuristicDistance(s, target) + heuristicDistance(s, below);
-                    return heuristicDifference(s, target, below, true);
-                }
-                // Same heuristic as for grabbing the target.
-                return heuristicDistance(s, target);
-
-            default:
-                throw new Planner.Error("!!! Unimplemented relation in heuristicAtom: "+atom.rel);
-                return 0;
-        }
-
-        return 0;
-    }
-
-    // Another attempt at heuristic function. Doesn't work as well as the simple
-    // grab heuristic. However, that one isnt really admissible...
-    function heuristicDifference(s : State, above : String, below : String, exactlyOntop : boolean) : number {
-
-        var stackA  : number;
-        var stackB  : number;
-        var heightA : number;
-        var heightB : number;
-        var aboveA  : number;
-        var aboveB  : number;
-
-        // TODO floor...
-
-        for(var stackNo in s.stacks){
-            var stack = s.stacks[stackNo];
-            for(var height in stack){
-                if(stack[height] === above){
-                    stackA = stackNo;
-                    heightA = height;
-                }
-                if(stack[height] === below){
-                    stackB = stackNo;
-                    heightB = height;
-                }
-            }
-        }
-
-        var holdCost = 0;
-        if(s.holding != null){
-            holdCost = 1;
-        }
-
-        // OBS, can be in arm as well...
-        if(s.holding === above){
-            stackA = s.arm;
-            aboveA = 0;
-        } else {
-            aboveA = s.stacks[stackA].length -1 -heightA;
-        }
-
-        if(s.holding === below){
-            stackB = s.arm;
-            aboveB = 0;
-        } else {
-            aboveB = s.stacks[stackB].length -1 -heightB;
-        }
-
-        var armCost = abs(s.arm - stackA) + abs(stackA - stackB);
-
-        // Number of objects that needs to be moved.
-        var aboveCost;
-        if(exactlyOntop){
-            // if(stackA === stackB){
-            //     aboveCost = 1 + max(aboveA, aboveB);
-            // } else {
-            //     aboveCost = aboveA + aboveB;
-            // }
-            aboveCost = aboveA + aboveB;
-        } else {
-            // ie Just somewhere above is sufficient
-            throw new Planner.Error("should not be here atm...");
-        }
-        // if(isUndefined(aboveCost)) throw new Planner.Error("aboveCost undefined!");
-        return holdCost + armCost + 4*aboveCost;
-    }
-
-
-
-    // Computes the expected number of actions to grab an object
-    function heuristicDistance(s : State, target : String) : number {
-        if(s.holding === target){
-            return 0;
-        }
-
-        var holdCost = 0;
-        if(s.holding != null){
-            holdCost = 1;
-        }
-
-        var emptyStacks = 0;
-        for(var stackNo in s.stacks){
-            if(s.stacks[stackNo].length == 0){
-                emptyStacks = emptyStacks +1;
-            }
-        }
-
-        if(target === "floor"){
-            var closest = Infinity;
-
-            for(var stackNo in s.stacks){
-                var stack = s.stacks[stackNo];
-                var cost = abs(stackNo - s.arm) + 4*stack.length;
-                if(closest > cost){
-                    closest = cost;
-                }
-            }
-            return closest;
-        }
-
-        for(var stackNo in s.stacks){
-            var stack = s.stacks[stackNo];
-            for(var height in stack){
-                if(stack[height] === target){
-                    var objectsAbove = stack.length -1 -height;
-                    return abs(stackNo - s.arm) + 4*objectsAbove + holdCost; // - 0.5*emptyStacks;
-                }
-            }
-        }
-
-        throw new Planner.Error("!!! Error in heuristicDistance: must be able to find the target somewhere in the world...");
-        // Of course unless the target is simply the floor...
-
-
-        return 0;
-
-    }
 
     /**
     * @return goal function for Astar.
     */
-    function computeGoalFunction(intprt : Interpreter.Literal[][]) : Astar.Goal<State>{
-        return (s : State) => {
+    function computeGoalFunction(intprt : Interpreter.Literal[][]) : Astar.Goal<Heuristics.State>{
+        return (s : Heuristics.State) => {
             for(var ix in intprt){
                 if(testConjunctiveClause(s, intprt[ix])){
                     return true;
@@ -268,7 +81,7 @@ module Planner {
         };
     }
 
-    function testConjunctiveClause(s : State, c : Interpreter.Literal[]) : boolean{
+    function testConjunctiveClause(s : Heuristics.State, c : Interpreter.Literal[]) : boolean{
         for(var ix in c){
             if(! testAtom(s, c[ix])){
                 return false;
@@ -278,7 +91,7 @@ module Planner {
         return true;
     }
 
-    function testAtom(s : State, atom : Interpreter.Literal) : boolean {
+    function testAtom(s : Heuristics.State, atom : Interpreter.Literal) : boolean {
         var ret = (result => {
             if(atom.pol){
                 return result;
@@ -323,7 +136,7 @@ module Planner {
         return ret(false);
     }
 
-    function neighbours(s : State) : Astar.Neighb<State>[]{
+    function neighbours(s : Heuristics.State) : Astar.Neighb<Heuristics.State>[]{
         var result = [];
         var numStacks = s.stacks.length;
 
@@ -359,7 +172,7 @@ module Planner {
         return result;
     }
 
-    function performAction(action: string, state: State) : Astar.Neighb<State>{
+    function performAction(action: string, state: Heuristics.State) : Astar.Neighb<Heuristics.State>{
         var newState = cloneState(state);
 
         switch(action){
@@ -397,21 +210,7 @@ module Planner {
         return typeof a === 'undefined' ;
     }
 
-    function max(a, b){
-        if(a > b){
-            return a;
-        }
-        return b;
-    }
-
-    function abs(a : number) : number{
-        if(a < 0){
-            return -a;
-        }
-        return a;
-    }
-
-    function cloneState(s : State) : State{
+    function cloneState(s : Heuristics.State) : Heuristics.State{
         var rs = [];
         for(var i in s.stacks){
             rs.push(s.stacks[i].slice());
